+++ qgreenland/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "v4.0.0alpha3"
+__version__ = "v4.0.0alpha4"